fix(settings): 修复 Token 统计大数被 Int 截断显示为负数
formatTokenCount 参数由 Int 改为 Long，去掉统计页总消耗、趋势图 Y 轴、渠道与模型排行、缓存副标题的 toInt() 截断，token 缩写新增 B 档。

+++ app/src/main/java/com/aicode/feature/agent/presentation/component/ChatHeaderComponents.kt
@@ -219,8 +219,8 @@ private fun ConnectionIndicator(
 
 @Composable
 private fun TokenStats(inputTokens: Int, outputTokens: Int) {
-    val inStr = formatTokenCount(inputTokens)
-    val outStr = formatTokenCount(outputTokens)
+    val inStr = formatTokenCount(inputTokens.toLong())
+    val outStr = formatTokenCount(outputTokens.toLong())
     Text(
         text = "↑$inStr ↓$outStr",
         style = MaterialTheme.typography.bodySmall,
